requirements.txt
bs4
pyenchant
requests
discord.py
python-dotenv
wikipedia-api
json
numpy
sympy
scipy
